Clearly organize text to pre-approval required if Per-Diem Lodging rates exceeded.
Based on discussion with Conservancy's Treasurer, I reworked this section
to clearly state that Conservancy can approve higher rates for hotels, but
that such approval should be done before travel begins.  I also cut down
some of the extra detail in this section that I think is obvious (i.e.,
how to use the GSA website).