Flight price search must be saved and submitted: now Clarified in main policy.
In the "Easy Reference Guide", it had been noted that you must include "Your
fare search (i.e., a screen shot)" when submitting a reimbursement request.
However, the main body of the text didn't explicitly say that.

I've also reintroduced herein a list of fare search sites.  We took out the
list initially because some complained that it prescribed a specific search
site.  The new wording herein makes it clear that the list is a proper subset
of acceptable search sites.  More can be added if folks like.