Outright reject requests after 90 days.
The "Board of Directors approval for 90 days late" rule was originally
instituted to make it clear that reimbursements beyond 90 days were
really difficult for Conservancy to process.  In practice, the rule has
meant that travelers simply believe that they "have another out" if they
are late, and late reimbursement requests put an undue burden on
Conservancy staff for special processing, and they create extra work for
our volunteer Board of Directors.

As such, Conservancy now simply rejects reimbursement requests that are
received 90 days or more after the last day of travel.